ContactRequestModule.java created (?)
This file is also not tested, but at least it compiles.

Note - when we get the req.param.value(), it returns as a string.
Because of this, we are needing to parseInt on the req.param.value
This is obviously terrible and i will look into a better way to approach this